+++ NEWS.rst
@@ -6,7 +6,10 @@ Lists of changes between versions.
 Latest
 ------
 * Minor: Use licensecheck instead of liccheck.
+<<<<<<< Updated upstream
 * Minor: Added type marker.
+=======
+>>>>>>> Stashed changes
 
 2.0.0
 ------