+++ NEWS.rst
@@ -5,7 +5,7 @@ Lists of changes between versions.
 
 Latest
 ------
-* Minor: Use brundle to run all linters.git dif
+* Minor: Use brundle to run all linters.
 
 2.1.0
 -----