+++ NEWS.rst
@@ -7,6 +7,7 @@ Latest
 ------
 * Minor: Added bouillon cli on install.
 * Minor: Added default_branch() to bouillon.git.
+* Minor: Support Python 3.13.
 * Minor: Use brundle to run all linters.
 * Minor: Dropped setup, lint, test and upgrade commands from example boil.py.